pyproject.toml
[project]
name = "productmind-ai"
version = "0.1.0"
description = "AI Product Manager agent for enterprise teams"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pinecone-client>=3.0.0",
    "upstash-redis>=1.2.0",
    "inngest>=0.4.0",
    "structlog>=24.4.0",
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-env>=1.1.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
env = [
    "APP_ENV=test",
    "DATABASE_URL=postgresql+asyncpg://postgres:postgres@localhost:5432/postgres_test",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
